foreclosure-agent/requirements.txt
playwright==1.49.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.1
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
